refactor: Extract shared test func
